Medicina/requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
python-dotenv==0.19.2
PyJWT==1.7.1
marshmallow==3.14.1
PyMySQL==1.0.2
flask-sqlalchemy==2.5.1